Fix bug hidden on CI and make ci failed if tests failure

+++ models/migrations/v1_22/v283.go
@@ -5,6 +5,7 @@ package v1_22 //nolint
 
 import (
 	"xorm.io/xorm"
+	"xorm.io/xorm/schemas"
 )
 
 func AddCombinedIndexToIssueUser(x *xorm.Engine) error {
@@ -20,8 +21,14 @@ func AddCombinedIndexToIssueUser(x *xorm.Engine) error {
 		return err
 	}
 	for _, issueUser := range duplicatedIssueUsers {
-		if _, err := x.Exec("delete from issue_user where id in (SELECT id FROM issue_user WHERE issue_id = ? and uid = ? limit ?)", issueUser.IssueID, issueUser.UID, issueUser.Cnt-1); err != nil {
-			return err
+		if x.Dialect().URI().DBType == schemas.MSSQL {
+			if _, err := x.Exec("delete from issue_user where id in (SELECT top ? id FROM issue_user WHERE issue_id = ? and uid = ?)", issueUser.Cnt-1, issueUser.IssueID, issueUser.UID); err != nil {
+				return err
+			}
+		} else {
+			if _, err := x.Exec("delete from issue_user where id in (SELECT id FROM issue_user WHERE issue_id = ? and uid = ? limit ?)", issueUser.IssueID, issueUser.UID, issueUser.Cnt-1); err != nil {
+				return err
+			}
 		}
 	}
